requirements.txt
# requirements.txt
streamlit
python-whois
requests
